move-to-firestore/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-firestore>=1.6
python-dateutil>=2.8
